Clear building selection after successful placement.
The selected building is cleared after successfully placing a building instead of when right clicking. This aligns better with the one building per turn design choice.

We will need to set up a flag that is checked prior to selection and is reset at the start/end of a turn.